+++ src/Vehicle.cpp
@@ -7,13 +7,9 @@
 #include <string>
 #include "Cost.h"
 
-/**
- * Initializes Vehicle
- */
-
 Vehicle::Vehicle() {}
 
-Vehicle::Vehicle(int lane, float s, float v, float a, string state)
+Vehicle::Vehicle(int lane, double s, double v, double a, string state)
 {
     this->lane = lane;
     this->s = s;
@@ -39,8 +35,8 @@ vector<Vehicle> Vehicle::choose_next_state(map<int, vector<Vehicle>> predictions
 
     */
     vector<string> states = successor_states();
-    float cost;
-    vector<float> costs;
+    double cost;
+    vector<double> costs;
     vector<string> final_states;
     vector<vector<Vehicle>> final_trajectories;
 
@@ -55,8 +51,8 @@ vector<Vehicle> Vehicle::choose_next_state(map<int, vector<Vehicle>> predictions
         }
     }
 
-    vector<float>::iterator best_cost = min_element(begin(costs), end(costs));
-    int best_idx = distance(begin(costs), best_cost);
+    vector<double>::iterator best_cost = min_element(begin(costs), end(costs));
+    int best_idx = static_cast<int>(distance(begin(costs), best_cost));
     return final_trajectories[best_idx];
 }
 
@@ -120,17 +116,18 @@ vector<Vehicle> Vehicle::generate_trajectory(string state, map<int, vector<Vehic
     return trajectory;
 }
 
-vector<float> Vehicle::get_kinematics(map<int, vector<Vehicle>> predictions, int lane)
+vector<double> Vehicle::get_kinematics(map<int, vector<Vehicle>> predictions, int lane)
 {
     /*
     Gets next timestep kinematics (position, velocity, acceleration)
     for a given lane. Tries to choose the maximum velocity and acceleration,
     given other vehicle positions and accel/velocity constraints.
     */
-    float max_velocity_accel_limit = this->max_acceleration + this->v;
-    float new_position;
-    float new_velocity;
-    float new_accel;
+    double max_velocity_accel_limit = this->max_acceleration + this->v;
+    double new_position;
+    double new_velocity;
+    double new_accel;
+
     Vehicle vehicle_ahead;
     Vehicle vehicle_behind;
 
@@ -143,7 +140,7 @@ vector<float> Vehicle::get_kinematics(map<int, vector<Vehicle>> predictions, int
         }
         else
         {
-            float max_velocity_in_front =
+            double max_velocity_in_front =
                 (vehicle_ahead.s - this->s - this->preferred_buffer) + vehicle_ahead.v - 0.5 * (this->a);
             new_velocity = min(min(max_velocity_in_front, max_velocity_accel_limit), this->target_speed);
         }
@@ -163,7 +160,7 @@ vector<Vehicle> Vehicle::constant_speed_trajectory()
     /*
     Generate a constant speed trajectory.
     */
-    float next_pos = position_at(1);
+    double next_pos = position_at(1);
     vector<Vehicle> trajectory = {Vehicle(this->lane, this->s, this->v, this->a, this->state),
                                   Vehicle(this->lane, next_pos, this->v, 0, this->state)};
     return trajectory;
@@ -175,10 +172,10 @@ vector<Vehicle> Vehicle::keep_lane_trajectory(map<int, vector<Vehicle>> predicti
     Generate a keep lane trajectory.
     */
     vector<Vehicle> trajectory = {Vehicle(lane, this->s, this->v, this->a, state)};
-    vector<float> kinematics = get_kinematics(predictions, this->lane);
-    float new_s = kinematics[0];
-    float new_v = kinematics[1];
-    float new_a = kinematics[2];
+    vector<double> kinematics = get_kinematics(predictions, this->lane);
+    double new_s = kinematics[0];
+    double new_v = kinematics[1];
+    double new_a = kinematics[2];
     trajectory.push_back(Vehicle(this->lane, new_s, new_v, new_a, "KL"));
     return trajectory;
 }
@@ -188,13 +185,13 @@ vector<Vehicle> Vehicle::prep_lane_change_trajectory(string state, map<int, vect
     /*
     Generate a trajectory preparing for a lane change.
     */
-    float new_s;
-    float new_v;
-    float new_a;
+    double new_s;
+    double new_v;
+    double new_a;
     Vehicle vehicle_behind;
     int new_lane = this->lane + lane_direction[state];
     vector<Vehicle> trajectory = {Vehicle(this->lane, this->s, this->v, this->a, this->state)};
-    vector<float> curr_lane_new_kinematics = get_kinematics(predictions, this->lane);
+    vector<double> curr_lane_new_kinematics = get_kinematics(predictions, this->lane);
 
     if (get_vehicle_behind(predictions, this->lane, vehicle_behind))
     {
@@ -205,8 +202,8 @@ vector<Vehicle> Vehicle::prep_lane_change_trajectory(string state, map<int, vect
     }
     else
     {
-        vector<float> best_kinematics;
-        vector<float> next_lane_new_kinematics = get_kinematics(predictions, new_lane);
+        vector<double> best_kinematics;
+        vector<double> next_lane_new_kinematics = get_kinematics(predictions, new_lane);
         // Choose kinematics with lowest velocity.
         if (next_lane_new_kinematics[1] < curr_lane_new_kinematics[1])
         {
@@ -244,7 +241,7 @@ vector<Vehicle> Vehicle::lane_change_trajectory(string state, map<int, vector<Ve
         }
     }
     trajectory.push_back(Vehicle(this->lane, this->s, this->v, this->a, this->state));
-    vector<float> kinematics = get_kinematics(predictions, new_lane);
+    vector<double> kinematics = get_kinematics(predictions, new_lane);
     trajectory.push_back(Vehicle(new_lane, kinematics[0], kinematics[1], kinematics[2], state));
     return trajectory;
 }
@@ -254,7 +251,7 @@ void Vehicle::increment(int dt = 1)
     this->s = position_at(dt);
 }
 
-float Vehicle::position_at(int t)
+double Vehicle::position_at(int t)
 {
     return this->s + this->v * t + this->a * t * t / 2.0;
 }
@@ -265,7 +262,7 @@ bool Vehicle::get_vehicle_behind(map<int, vector<Vehicle>> predictions, int lane
     Returns a true if a vehicle is found behind the current vehicle, false otherwise. The passed reference
     rVehicle is updated if a vehicle is found.
     */
-    int max_s = -1;
+    double max_s = -1.0;
     bool found_vehicle = false;
     Vehicle temp_vehicle;
     for (map<int, vector<Vehicle>>::iterator it = predictions.begin(); it != predictions.end(); ++it)
@@ -287,7 +284,7 @@ bool Vehicle::get_vehicle_ahead(map<int, vector<Vehicle>> predictions, int lane,
     Returns a true if a vehicle is found ahead of the current vehicle, false otherwise. The passed reference
     rVehicle is updated if a vehicle is found.
     */
-    int min_s = this->goal_s;
+    double min_s = this->goal_s;
     bool found_vehicle = false;
     Vehicle temp_vehicle;
     for (map<int, vector<Vehicle>>::iterator it = predictions.begin(); it != predictions.end(); ++it)
@@ -312,8 +309,8 @@ vector<Vehicle> Vehicle::generate_predictions(int horizon)
     vector<Vehicle> predictions;
     for (int i = 0; i < horizon; i++)
     {
-        float next_s = position_at(i);
-        float next_v = 0;
+        double next_s = position_at(i);
+        double next_v = 0;
         if (i < horizon - 1)
         {
             next_v = position_at(i + 1) - s;
